3.1.7	21/02/2024 Bug Fixed: Get header pada tool image, Tidak bisa add new product

+++ willowmgr12/model/catalog/product.php
@@ -166,7 +166,7 @@ class ModelCatalogProduct extends Model
 				}
 			}
 		}
-		
+
 		# Product Variant
 		$this->db->query("DELETE FROM " . DB_PREFIX . "product_option_value WHERE product_id = '" . (int)$product_id . "'");
 
@@ -364,27 +364,28 @@ class ModelCatalogProduct extends Model
 		$product_variant = $this->getProductVariants($product_id);
 
 		$images = array_merge($images, array_column($product_variant['variant'], 'image'));
+		var_dump($images);
 
 		foreach ($images as $image) {
-			$extension = pathinfo(DIR_IMAGE . $image, PATHINFO_EXTENSION);
+			if (!empty($image)) {
+				$extension = pathinfo(DIR_IMAGE . $image, PATHINFO_EXTENSION);
 
-			$cache_files = str_replace('.' . $extension, '*', DIR_IMAGE . 'cache/' . $image);
+				$cache_files = str_replace('.' . $extension, '*', DIR_IMAGE . 'cache/' . $image);
 
-			foreach (glob($cache_files) as $file) {
-				if (file_exists($file)) {
-					unlink($file);
-					var_dump($image);
+				foreach (glob($cache_files) as $file) {
+					if (file_exists($file)) {
+						unlink($file);
+						var_dump($image);
+					}
 				}
-			}
 
-			if (file_exists(DIR_IMAGE . $image)) {
-				unlink(DIR_IMAGE . $image);
-				var_dump($image);
+				if (file_exists(DIR_IMAGE . $image)) {
+					unlink(DIR_IMAGE . $image);
+				}
 			}
 		}
 
 		clearstatcache();
-		die('---breakpoint---');
 
 		$this->db->query("DELETE FROM " . DB_PREFIX . "product WHERE product_id = '" . (int)$product_id . "'");
 		$this->db->query("DELETE FROM " . DB_PREFIX . "product_attribute WHERE product_id = '" . (int)$product_id . "'");
@@ -493,7 +494,7 @@ class ModelCatalogProduct extends Model
 				$filter_data[] = "ps.product_special_id IS NOT NULL";
 			} elseif ($data['filter_special'] == 'discount') {
 				$join_data[] = "LEFT JOIN `oc_product_discount` pd2 ON (pd2.product_id = p.product_id)";
-			
+
 				$filter_data[] = "pd2.product_discount_id IS NOT NULL";
 			}
 		}
@@ -964,7 +965,7 @@ class ModelCatalogProduct extends Model
 				$filter_data[] = "ps.product_special_id IS NOT NULL";
 			} elseif ($data['filter_special'] == 'discount') {
 				$join_data[] = "LEFT JOIN `oc_product_discount` pd2 ON (pd2.product_id = p.product_id)";
-			
+
 				$filter_data[] = "pd2.product_discount_id IS NOT NULL";
 			}
 		}
